refactor(viewer): extract viewer_home.jsx from viewer.jsx
Pure file split, no behavior change. Step 10 of mp-pxxc.
