+++ web-mobile/index.html
@@ -108,6 +108,10 @@
          extracted from viewer.jsx (mp-pxxc step 9).
          Must load after viewer_schedule.js and before viewer.js. -->
     <script type="module" src="/dist/viewer_competition.js?v=5"></script>
+    <!-- Home page component + routing helpers (ViewerHome, shouldShowRegister, resolveDeepLink),
+         extracted from viewer.jsx (mp-pxxc step 10).
+         Must load after viewer_competition.js and before viewer.js. -->
+    <script type="module" src="/dist/viewer_home.js?v=5"></script>
     <script type="module" src="/dist/viewer.js?v=5"></script>
     <script type="module" src="/dist/admin_helpers.js?v=5"></script>
     <script type="module" src="/dist/admin_shell.js?v=5"></script>
